ioctl fix

+++ exercise_33/message_slot.h
@@ -9,7 +9,7 @@
 #define MINOR_NUM
 // Set the message of the device driver
 //#define IOCTL_MSG_SLOT_CHANNEL _IOW(MAJOR_NUM, 0, unsigned int)//TODO: change the 0 to MINOR_NUM
-#define IOCTL_MSG_SLOT_CHANNEL _IOW(MAJOR_NUM, 0, unsigned long)
+#define IOCTL_MSG_SLOT_CHANNEL _IOW(MAJOR_NUM, 0, unsigned int)
 
 #define DEVICE_RANGE_NAME "char_dev" // TODO: check this line
 #define BUF_LEN 128
